listing/go1-bmsMonitor/bmsMonitor/requirements.txt
paho-mqtt ~= 1.6